rename `ast` to `sat`, since it's a mid_term after parsing instead of analyzing
